clase_admin/requirements.txt
asgiref==3.8.1
Django==5.1
sqlparse==0.5.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-import-export==4.1.1
django-rosetta==0.10.1
django-modeltranslation==0.19.9
mysqlclient==2.2.4